Lesson 01: Pull up ronding to integer type using a constexpr function.
Resulting of our group's review, we concluded this increases readability.

We also realized that implementing the `dx` (x-axis distance)
and `sx` parameter (x-axis scaling) distracts from the original task.
However, these are the costs to pay to move to a more 'production like' code.
Its application is shown within the `main` function of printing all objects
in all sizes aligned to their center.