Review of STL iterator + difference_type limitations (half range) compared to size_t index ops and pointer arithmetic
Svenson an I discussed the issue and we were a bit puzzled ;-)

So it would have been great to have size_type to be unsigned (ssize_t)
to simply constrain and harmonize the whole std::vector (etc) operations.

Alternatively, an iterator could have been a pointer type,
where such operations are at least tolerated
despite similar considerations to ptrdiff_t like
	void *lo, *hi; lo-hi !

In the meantime, if using iterator and a size,
stick to iterator and signed size.